requirements.txt
arrow==0.7.0
Django==1.9.7
django-bootstrap3==7.0.1
django-cache-url==1.0.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.7
django-rq==0.9.1
django-rq-jobs==0.1.5.1
django-smart-selects==1.2.2
gunicorn==19.6.0
ipdb==0.10.0
ipython==4.2.0
mysqlclient==1.3.7
psycopg2==2.6.1
pylibmc==1.5.1
python-dateutil==2.5.3
redis==2.10.5
rq==0.6.0
rq-scheduler==0.6.1
six==1.10.0
watchdog==0.8.3
Werkzeug==0.11.10
-e git+https://github.com/wk-tech/python-smsfly@v0.1.2#egg=smsfly
